pyproject.toml
[project]
name = "pfeiffer-turbo"
version = "0.2.0"
description = ""
authors = [{ name = "ograsdijk", email = "o.grasdijk@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
	"pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
	"pytest>=8.0",
	"ruff>=0.5",
	"mypy>=1.8",
]

[project.urls]
Repository = "https://github.com/ograsdijk/pfeiffer-turbo"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pfeiffer_turbo*"]

[tool.setuptools.package-data]
pfeiffer_turbo = ["py.typed"]
